# gdpd-backend/gdpd-pedidos-api/src/main/java/com/bbva/gdpd/gdpdpedidosapi/dto/CreatePedidoRequest.java
package com.bbva.gdpd.gdpdpedidosapi.dto;

import javax.validation.constraints.DecimalMin;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;
import java.math.BigDecimal;

public class CreatePedidoRequest {
    @NotBlank(message = "La referencia es obligatoria")
    private String referencia;
    @NotNull(message = "El importe es obligatorio")
    @DecimalMin(value = "0.01", message = "El importe debe ser mayor que 0")
    private BigDecimal importe;
    public String getReferencia() { return referencia; }
    public void setReferencia(String referencia) { this.referencia = referencia; }
    public BigDecimal getImporte() { return importe; }
    public void setImporte(BigDecimal importe) { this.importe = importe; }
}