# gdpd-backend/gdpd-pedidos-api/src/main/resources/openapi/api.yml
swagger: '2.0'
info:
  version: '1.0.0'
  title: GDPD Pedidos API
  description: API REST para gestión del ciclo de vida de pedidos en NOVA COSMOS BBVA
  contact:
    url: http://www.bbva.com
    name: BBVA GDPD Team
    email: gdpd@bbva.com
host: localhost:8080
x-generator-properties:
  business-unit: gdpd
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /pedidos:
    get:
      summary: Listar pedidos
      operationId: getPedidos
      tags: [Pedidos]
      parameters:
        - name: estado
          in: query
          required: false
          type: string
          enum: [PENDIENTE, EN_PROCESO, COMPLETADO, CANCELADO]
      responses:
        '200':
          description: Lista de pedidos
          schema:
            type: array
            items:
              $ref: '#/definitions/PedidoDTO'
    post:
      summary: Crear pedido
      operationId: createPedido
      tags: [Pedidos]
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreatePedidoRequest'
      responses:
        '201':
          description: Pedido creado
          schema:
            $ref: '#/definitions/PedidoDTO'
  /pedidos/{id}:
    get:
      summary: Obtener pedido por ID
      operationId: getPedidoById
      tags: [Pedidos]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Pedido encontrado
          schema:
            $ref: '#/definitions/PedidoDTO'
        '404':
          description: No encontrado
    put:
      summary: Actualizar pedido
      operationId: updatePedido
      tags: [Pedidos]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdatePedidoRequest'
      responses:
        '200':
          description: Pedido actualizado
          schema:
            $ref: '#/definitions/PedidoDTO'
    delete:
      summary: Eliminar pedido
      operationId: deletePedido
      tags: [Pedidos]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: Eliminado
  /pedidos/events:
    get:
      summary: Stream SSE de eventos de pedidos
      operationId: streamPedidoEvents
      tags: [Pedidos SSE]
      produces:
        - text/event-stream
      responses:
        '200':
          description: Stream de eventos
definitions:
  PedidoDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
      referencia:
        type: string
      estado:
        type: string
        enum: [PENDIENTE, EN_PROCESO, COMPLETADO, CANCELADO]
      importe:
        type: number
        format: double
      fechaCreacion:
        type: string
        format: date-time
  CreatePedidoRequest:
    type: object
    required: [referencia, importe]
    properties:
      referencia:
        type: string
      importe:
        type: number
        format: double
  UpdatePedidoRequest:
    type: object
    properties:
      estado:
        type: string
        enum: [PENDIENTE, EN_PROCESO, COMPLETADO, CANCELADO]
      importe:
        type: number
        format: double